Fix import for notifier
